copy the last non-diacritic's node when inserting kashida
A kashida is inserted between the glyphs of two non-diacritic
characters. Some extra logic is required when a diacritic is also
assigned to the character before the kashida. E.g. TEH FATHA
THEH (where FATHA is a diacritic), which should be transformed to TEH
FATHA KASHIDA THEH. In this case the diacritic is to be the
predecessor of the kashida. However, the preceding non-diacritic
character is the kashida's reference glyph, which it has to adapt to.

When inserting the kashida, a new node is created by copying an
existing node. It's critical, to copy the right node.

If we would copy the node of the diacritic to make a new node for the
kashida, the node would have the x and y offsets of the diacritic,
which leads to badly shifted kashidas in the results. Therefore, we
keep track of the last non-diacritic character which is the real
referencial character for the kashida. It's node is copied to get the
correct offsets for the kashida.