added a logging device for inspecting predecessors of kashida
Input characters seem to be replaced with ligatures before kashida
insertion is done. This makes setting up non-standard hyphenation
rules hardly manageable unless we have an device for
inspectation. Adding the following code after loading the babel
package results in a log message with the unicode codepoint of the
predecessors of inserted kashidas.

\directlua{Babel.arabic.verbose = true}