added an option to turn off kashidas after ligatures as asked in #258
Replacing simple characters with ligatures seems to be done before
inserting kashidas. This may result in incorrect kashida insertion.

If we have e.g. letters BEH and YEH where a kashida can be inserted in
between. But we also have an BEH+YEH ligature. Now, the glyph on the
a's position gets a kashida weight, but it has been replaced with the
BEH+YEH ligature. So, the result will be inserted kashidas after the
ligature, i.e. after YEH, which is not what we want.

Test case: {بِيَ} with Arabic Typesetting