allow to change the reference glyph for kashidas inserted after diacritics
When a kashida is inserted behind a diacritics, the diacritic's glyph
is not the node which should be used for creating the kashida's
node (by copying), but the character's node. Copying the diacritic's
node results in x and y offset, which displace the kashida. By copying
the node of the character (on which the diacritic is placed) or by
copying a ghost node we get correct x and y offsets to adapt the
kashida smoothly to the character.

This feature can be enabled by
`\directlua{Babel.arabic.kashida_after_diacritics=true}` or disabled
by setting it to `false`.

This fixes issue #257. But setting
`Babel.arabic.kashida_after_diacritics` from the name of the
transformation rules would be great.